requirements.txt
absl-py==2.2.2
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.1
contourpy==1.1.1
coremltools==8.3.0
cycler==0.12.1
exceptiongroup==1.2.2
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.4.0
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
joblib==1.4.2
keras==2.11.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
numpy==1.24.4
oauthlib==3.2.2
onnx==1.12.0
onnxruntime-gpu @ file:///ultralytics/onnxruntime_gpu-1.8.0-cp38-cp38-linux_aarch64.whl
opencv-python==4.11.0.86
openvino==2024.0.0
openvino-telemetry==2025.1.0
opt_einsum==3.4.0
packaging==20.9
pandas==2.0.3
pillow==10.4.0
protobuf==3.19.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyaml==25.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-cpu-aws==2.11.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.35.0
tensorflowjs==3.18.0
tensorrt @ file:///ultralytics/tensorrt-8.2.0.6-cp38-none-linux_aarch64.whl
termcolor==2.4.0
threadpoolctl==3.5.0
torch @ https://github.com/ultralytics/assets/releases/download/v0.0.0/torch-1.11.0a0+gitbc2c6ed-cp38-cp38-linux_aarch64.whl
torchvision @ https://github.com/ultralytics/assets/releases/download/v0.0.0/torchvision-0.12.0a0+9b5a3fe-cp38-cp38-linux_aarch64.whl
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
-e git+https://github.com/ultralytics/ultralytics@efdafa9c451fbd4dcf77977adccb367b277c1913#egg=ultralytics
ultralytics-thop==2.0.14
urllib3==2.2.3
uv==0.7.0
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
